// src/types/ExtendedTheme.d.ts
export interface ExtendedTheme extends Theme {
    dark: boolean,
    colors: {
      primary: string,
      background: string,
      card: string,
      text: string,
      border: string,
      notification: string,
      red: string,
      green: string,
      yellow: string,
      secondaryCard: string,
    }
  }